switch out deprecated docker client library
closes #2071
